requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
aiosqlite>=0.19.0
Pillow>=10.0.0
requests>=2.31.0